unify current_user and authentications since they share concerns
