+++ app/controllers/application_controller.rb
@@ -6,10 +6,7 @@ class ApplicationController < ActionController::Base
   # Force all access to the app over SSL, use Strict-Transport-Security, and use secure cookies.
   force_ssl if: :force_ssl?
 
-  # CurrentUser must be after protect_from_forgery,
-  # so that authenticate! is called
-  include CurrentUser
-  include Authorization
+  include CurrentUser # must be after protect_from_forgery, so that authenticate! is called
 
   protected
