+++ app/controllers/concerns/project_level_authorization.rb
@@ -1,6 +1,5 @@
 module ProjectLevelAuthorization
   extend ActiveSupport::Concern
-  include Authorization
   include CurrentProject
 
   private
