+++ test/controllers/concerns/current_user_test.rb
@@ -14,10 +14,36 @@ describe "CurrentUser included in controller" do
       Stage.find(params[:id]).update_attribute(:name, 'MUUUU')
       head :ok
     end
+
+    def super_admin_action
+      authorize_super_admin!
+      head :ok
+    end
+
+    def admin_action
+      authorize_admin!
+      head :ok
+    end
+
+    def deployer_action
+      authorize_deployer!
+      head :ok
+    end
+
+    def unauthorized_action
+      unauthorized!
+    end
   end
 
   tests CurrentUserTestController
   use_test_routes
+
+  def self.authorized(method, action, params)
+    it "is authorized to #{method} #{action}" do
+      send method, action, params
+      refute_unauthorized
+    end
+  end
 
   as_a_viewer do
     around { |t| PaperTrail.with_whodunnit(nil, &t) }
@@ -57,6 +83,38 @@ describe "CurrentUser included in controller" do
         @controller.send(:current_user).must_equal nil
         session.inspect.must_equal({}.inspect)
       end
+    end
+  end
+
+  as_a_viewer do
+    unauthorized :get, :deployer_action, test_route: true
+    unauthorized :get, :admin_action, test_route: true
+    unauthorized :get, :super_admin_action, test_route: true
+  end
+
+  as_a_deployer do
+    authorized :get, :deployer_action, test_route: true
+    unauthorized :get, :admin_action, test_route: true
+    unauthorized :get, :super_admin_action, test_route: true
+  end
+
+  as_a_admin do
+    authorized :get, :deployer_action, test_route: true
+    authorized :get, :admin_action, test_route: true
+    unauthorized :get, :super_admin_action, test_route: true
+  end
+
+  as_a_super_admin do
+    authorized :get, :deployer_action, test_route: true
+    authorized :get, :admin_action, test_route: true
+    authorized :get, :super_admin_action, test_route: true
+  end
+
+  as_a_viewer do
+    it "logs unautorized so we can see it in test output for easy debugging" do
+      Rails.logger.expects(:warn)
+      get :unauthorized_action, test_route: true
+      assert_unauthorized
     end
   end
